requirements.txt
requests == 2.32.3
pytest == 8.3.2
jsonschema == 4.23.0
allure-pytest == 2.13.5
requests == 2.32.3
python-dotenv == 1.0.1